diagnostics: rename diagnosticsd to wilco_dtc_supportd
The diagnosticsd does not manage diagnostics anymore and has to be
renamed to wilco_dtc_supportd since it is used only for Wilco project
interfaces.

The renaming is planned to be done in several stages:
  Rename the daemons: diagnosticsd->wilco_dtc_supportd and
  diagnostics_processor -> wilco_dtc.
  Rename the source code namespaces/classes/objects.
  Rename interfaces.

In scope of this CL only the diagnostics daemon and
diagnostics_processor daemon are renamed to wilco_dtc_support daemon.

Note: the occurrences of "diagnosticsd" and "diagnostics_processor" in comments
will be fixed in scope of the following CLs.

BUG=b:124423615
TEST=unittests
CQ-DEPEND=CL:1491351
CQ-DEPEND=CL:1479510

Change-Id: I0f791cf3e191101d4282038ca551541984e76bf4
Reviewed-on: https://chromium-review.googlesource.com/1479610
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Polina Bondarenko <pbond@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>