vm_tools: 9s: add support for using existing domain socket
We may want to avoid poking a hole in seccomp filter to be able to change
ownership of a newly created unix domain socket, and so may want to
listen on a pre-created socket that is passed to us by the parent
process.

Let's handle addresses in form of "unix-fd:NNN" to mean that
we should be using specified FD as a pre-existing socket.

BUG=b:123602808
TEST=Use seneschal_client to start 9P with given socket path, see that
data can be accessed by crosvm user.

Change-Id: I0ecca1137e580fbc0996f70c0de2ad1b097d6874
Reviewed-on: https://chromium-review.googlesource.com/1484433
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>