shill: wifi: Stop assuming kFailureOutOfRange as default disconnect reason
Currently, unexpected wifi disconnects are assumed to either be due to
bad credentials or to the AP being out of range. This change uses
disconnect reasons and signal strength in order to determine the the
proper disconnect reason to return.

One potential pitfall with the current logic is that SuspectOutOfRange
uses the last known BSS signal strength, which it polls from the AP
every 20 seconds. A signal strength below the threshold will lead to a
suspected out of range if no other reason presents itself. Thus it is
possible for a STA to move quickly away from the AP, such that the last
known signal strength is above the threshold, leading shill to determine
that the disconnect was caused due to unknown reasons instead of an out
of range error.

BUG=chromium:625387
TEST=- All unit tests are passing
     - Manually tested on chell chromebook. Walked around campus to
       generate legitimate out-of-range errors.
     - Tried connecting to SSID that didn't exist in the area, and
       ensured that Shill returned an "out-of-range" error (as the
       SSID presumably is available somewhere, but not here).

Change-Id: If88edc25078c48b3023f850990769ca9d9fa5e45
Reviewed-on: https://chromium-review.googlesource.com/1487574
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>