cryptohome: Support either-of-PCRs for TPM1.2 signature sealing
Update the TPM1.2 signature sealing backend to support restrictions
"the PCRs should be in either of the specified states", e.g.
"(PCRx==foo && PCRy==bar) OR (PCRx==bar && PCRz==baz)".

This paves the way for supporting the feature "allow login only when
specified PCRs are either in the default or in the user-specific
states" for challenge-response protected cryptohomes (a.k.a. smart
card login).

BUG=chromium:806788,b:77799573
TEST=existing tpm_live_test SignatureSealedSecretTest

Change-Id: I023eebb56611f642ddc76dc9e2fb1530e09e44cb
Reviewed-on: https://chromium-review.googlesource.com/1481031
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>