u2fd: Support software keys for U2F_REGISTER attestation.
A portion of a valid U2F_REGISTER response must be
signed by an attestation key, and the response must also
include an X509 certificate for that attestation key.

Currently cr50 provides this functionality, however for
non-G2F authentication this is not necessary, and the
attestation can be done purely in software.

This change adds support to u2fd for software-based
attestation.

BUG=b:123161715
TEST=manual tests on soraka

Change-Id: I9517f60f32426505890f7788dff5a091f52026de
Reviewed-on: https://chromium-review.googlesource.com/1486191
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>