shill: key_value_store: Change Convert* methods' output params to return values
The KeyValueStore::Convert* methods previously used an output parameter to
return their result--rather than simply returning the value--despite the
fact that a valid output value is always produced. This also allows for a
potential ambiguity of the interface in the case where the output parameter
points to a non-empty map. This change modifies the methods to simply return
the desired value.

BUG=chromium:517618
TEST=All unit tests are passing.

Change-Id: I769db7d28fc0b94c4bfa14d9ee2d2d3f251018a9
Reviewed-on: https://chromium-review.googlesource.com/1488251
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>