shill: Avoid map::emplace/insert on maps with exising keys
std::map::emplace and std::map::insert have the unintuitive behavior
that if the key already exists in the map, its value will not be
updated. When a map is created right before emplace/insert is called,
this does not cause any confusion. However, when updating a member
variable or an otherwise pre-existing map instance, it is best to
simply use operator[] (provided that the value type is
DefaultConstructible) unless we explicitly don't want to modify
existing key-value pairs.

BUG=chromium:517618
TEST=All unit tests are passing.

Change-Id: Ib30f285f0034ab2e082c9744291461c1a791f602
Reviewed-on: https://chromium-review.googlesource.com/1489992
Commit-Ready: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>