cryptohome: LockedToSingleUser file used to get the right keyset in decrypt flow.
At user login with password, if the keyset is bound to PCR, cryptohome
needs to know if the PCR was extended. To keep track of this, we create
a file in temporary storage when PCR is extended and use the existence of
the file to know which keyset needs to be used to decrypt. The file will
disappear automatically when device reboots.

BUG=b:77799573
TEST=None
CQ-DEPEND=CL:1496853

Change-Id: If3af9f3489c6eddf76922809b4bb4f70d8be57b7
Reviewed-on: https://chromium-review.googlesource.com/1489311
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>