cryptohome: Add salt into SignatureChallengeInfo
Add a new field into the SignatureChallengeInfo proto that
will hold the value of the salt used for challenge-protected vault
keysets. The |salt| field of the enclosing SerializedVaultKeyset
proto shouldn't be used for this purpose.

This allows to clearly separate in code the data flows of the
old-style salt (which is used for the cryptohome-TPM-key-based
encryption) and of the challenge-response salt.

This is a pure refactoring CL, no functional change is expected.

BUG=chromium:842791
TEST=none

Change-Id: I92f2765ea15569c3b7283532faece819d4e2cf8f
Reviewed-on: https://chromium-review.googlesource.com/1489591
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>