libbrillo: http: Allow for specification of custom certificate path
The HttpTransport classes use a default CA certificate path for all
communications. Specifying a custom certificate path is useful in
cases where we need to add a new set of certificates for a project,
but we don't want the certificates to be used by any other
communications.

BUG=chromium:847619
TEST=All unit tests are passing.

Change-Id: I4246d4acc526dde164a79a1b0afaf0c6ad409404
Reviewed-on: https://chromium-review.googlesource.com/1490871
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Alex Khouderchah <akhouderchah@chromium.org>