shill: simplfy Nl80211Frame::ToString()
Nl80211Frame::ToString() returns a boolean and appends the result to the
provided std::string* argument. The only scenario that it returns false
is when the std::string* argument is null. Instead, this CL simplifies
Nl80211Frame::ToString() to return a std::string and let the caller to
append the returned string to another std::string if needed.

BUG=None
TEST=Run unit tests.

Change-Id: I36afd99f2648db26335e23ed9e3723f2e8e4986a
Reviewed-on: https://chromium-review.googlesource.com/1493748
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>