shill: ethernet: Merge EthernetService constructors
In CL:1475043, we added a new EthernetService constructor to be used
by the generic ethernet_any service, which would not have a device
associated with it. Since all EthernetServices either have a device
or a storage_id_ set (but not both), we move these two fields into a
Properties struct.

BUG=chromium:933904
TEST=`FEATURES=test emerge-caroline shill` and
     `FEATURES=test emerge-whirlwind shill`

Change-Id: I3d6a64c64720923e62d9fabd47b65aeb0052c047
Reviewed-on: https://chromium-review.googlesource.com/1496426
Commit-Ready: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>