cryptohome: Pass whole KeyData to challenge-response code
Change the ChallengeCredentialsHelper's methods and related to
receive the whole KeyData instead of the smaller
ChallengePublicKeyInfo.

This allows to simplify the to-be-implemented callers by assuming
that the UsernamePasskey instances that are returned from the
ChallengeCredentialsHelper have already all key information
filled in. (As opposed to the callers having to overwrite the
KeyData with the current mount authorization key, hoping that no
useful field gets overwritten).

BUG=chromium:842791
TEST=ChallengeCredentials*Test

Change-Id: I2f7c3ca56c7def6d3a537754f730f1b05e53ba67
Reviewed-on: https://chromium-review.googlesource.com/1499433
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>