diagnostics: rename diagnosticsd to wilco_dtc_supportd
The diagnosticsd does not manage diagnostics anymore and has to be
renamed to wilco_dtc_supportd since it is used only for Wilco project
interfaces.

In scope of this CL only the files/classes/objects are renamed.

DPSL, mojom and DBus -related files/classes/objects/namespaces will be
renamed in scope of following CLs.

BUG=b:124423615
TEST=unittests

Change-Id: I896589dab4dd241307d046a02090c9c3c1c63d56
Reviewed-on: https://chromium-review.googlesource.com/1499573
Commit-Ready: Polina Bondarenko <pbond@chromium.org>
Tested-by: Polina Bondarenko <pbond@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>