shill: remove JSON store implementation
CL:293565 introduced a JSON-based on-disk format as an alternative to
the INI-like format for the shill profiles. The JSON-based format was
meant for Brillo and never became a replacement for the INI-like format
on Chrome OS. The two formats are currently mutually exclusive,
controlled by a 'json_store' USE flag, in the current implementation of
shill. If we later decide to migrate the INI-like format on Chrome OS to
an alternative format, we'd prefer protobuf over JSON as the former
offers a few practical advantages.

BUG=chromium:938587
TEST=Run unit tests.

Change-Id: Ia993b4832cb694a5af82a504ce43edf01aeaf80b
Reviewed-on: https://chromium-review.googlesource.com/1503794
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>