cryptohome: Fix memory leak in SignatureSealingBackendTpm1Impl
The GenerateCmk() function was leaking two buffers allocated by
Trousers due to wrongly specifying the TSS_HTPM handle instead of
the TSS_HCONTEXT one. This led to the construction of ScopedTssMemory
with a non-existant context handle, so that the consequent call of
Tspi_Context_FreeMemory() was a no-op.

BUG=chromium:806788
TEST=manual (log all allocations and deallocations in Trousers)

Change-Id: Ibb34278fd436b0f238a5dc9b8104d21a2c2a2500
Reviewed-on: https://chromium-review.googlesource.com/1504572
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>