arc-networkd: vm_tools: Generalize SubnetPool and add manager for guest networks.
This patch is the first step towards shared address provisioning and management
for guest networks. SubnetPool is generalized and the VM and container specific
aspects are removed. AddressManager is introduced to aggregate multiple subnet
pool that partition the Chrome OS private network space.

Concierge is updated to use the address manager to obtain VM, container and
plugin subnets.

arc-networkd will be updated in a subsequent patch to use this utility for
obtaining IPv4 addresses for arc0 and multinet.

BUG=b:77293260
TEST=ran presubmit for nocturne that includes: arc-networkd, shill and
vm_host_tools unit tests, and various Tast tests.
TEST=manual verification on flashed nocturne.

Change-Id: Ic7f2ea4b3d7c3fb3531027c34e6f23fe41f198be
Reviewed-on: https://chromium-review.googlesource.com/1506993
Commit-Ready: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>