attestation: provide RSA pub EK encrypted quote for EID computation
The quote for an RSA pub EK is computed when preparing for enrollment.
It is then encrypted with the public key of the Google CA that will
process the enrollment when the enrollment request is created.

The quote is only provided if the identity has a feature for
enrollment IDs, and if the main endorsement credential is not already
in RSA form.

BUG=chromium:939257
TEST=unit tests
CQ-DEPEND=CL:1487577

Change-Id: I0771fdc9bb05a58322628cb0e25306d5d124fe15
Reviewed-on: https://chromium-review.googlesource.com/1507694
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>