power: refactor RTC wakealarm usage model
Backdrop:
We used to have many false positive cases in partners' engineering
that the native suspend_stress_script been interrupted due to an
early wakeup of wakealarm.

With regard to the Chrome-OS power manager, we have some special power
manager test clients in native and remote which are designed to validate
no kernel/firmware issue during suspend path. These clients are basically
composed by the RTC wakealarm as the engine to perform test iterations,
typically these clients are not used by the end users.

Now there is a problem in between test clients and power manager daemon:
 - The test clients program RTC wakealarm before issuing suspend request
   through the dbus to the daemon
 - The power manager daemon has internal suspender delay policy allows
   registered applications to be ready for system suspending

Consider the interface is dbus, the above two are conflict because they do
not know each other in terms of encapsulation:
 - The test client would not know the daemon has delay policy while it
   programmed wakealarm first, i.e, suspender is opaque to test client
 - The power daemon policy instance has no idea w.r.t wakealarm budget
   it should follow-up, nor it has permission to access the RTC

Solution:
This patch refactors the RTC wakealarm assignment from before the
suspender to after it. Such that, the suspender delay policy will
not conflict w/ pre-programming the wakealarm. Also the wakealarm
timeout would approach the time we expect that the machine should
stay in a low power state.

In short explanation, it defers the wakealarm programming from the test
client to the powerd setuid helper script.

Note:
There is limitation for manual issuing the powerd_dbus_suspend w/
wakeup_timout, it requires a correct wakeup_count option as well.

BUG=chromium:941846, b:112335777, b:113242596, b:123243506, b:73018133
TEST=run the 'suspend_stress_test' w/ or w/o --memory_check option, run
the 'powerd_dbus_suspend' w/ or w/o wakeup_timeout respectively, examime
the footprints in powerd log and syslog to justify the wakealarm usage
against the Suspender.

Change-Id: I0fa3e11162ab8ca5565a342cdd7b098ee04e6947
Signed-off-by: Harry Pan <harry.pan@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1517064
Reviewed-by: Dan Erat <derat@chromium.org>