debugd: Return true when doing redudant coredump setting
There's been crashes in debugd in GetDBusError, and the only reason
those seem possible is by the ErrorPtr being null. I looked through all
the DBus method implementations in debugd, and this was the only case I
found where a method returns false but does not set the ErrorPtr to an
object. Rather than having these return an error (since this case is
just a redundant call, which we log) just have them return true which
should fix any crashes as a result of these calls.

BUG=chromium:939611
TEST=crosh upload_devcoredumps enable|disable works when ran repeatedly
and debugd no longer crashes from it
Change-Id: Ifad1cb4ce92bf791a3b3cfa5476ab88c19bef31a
Reviewed-on: https://chromium-review.googlesource.com/1511820
Commit-Ready: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>