power: Do not open sensor event files that won't be used
powerd preferences allow specifying that a proximity sensor will
not actually be used as a source of events to control transmit
power. The code, however, ends up opening the sensor event file
and holding on to it, even if downstream clients then do not
set up listening for events from that file. This prevents some
testing scenarios where raw events are read from the file for
validation (e.g. in a factory/lab situation).

This patch changes powerd to avoid opening sensor event files
altogether if it can detect that they will not be actually used.

BUG=b:113243504
TEST=added test cases, and observed no test regressions

Change-Id: I2efe6314341948f51c5d745b70443be8d2a9f3b5
Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1510514