kerberos: Implement account manager
Adds a way to set Kerberos configuration and implements all methods with
the exception of AutoConfig, which will have to use Samba to retrieve
the KDC IP (this is going to be a bit of work). As of now, the config
has to be provided manually.

Also hooks up the account manager with the Krb5Interface to acquire
Kerberos tickets.

BUG=chromium:932075
TEST=Emerges

Change-Id: I100ff5342d002c860e5b8583f47b59112631e4e3
Reviewed-on: https://chromium-review.googlesource.com/1475397
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>