cryptohome: fix IsAttestaionPrepared for tpm2
After CL:1418984, TpmIsAttestationPrepared in cryptohomed uses
GetEnrollmentPreparations to get "is prepared" status in the "distributed"
(TPM 2.0) case.

Before this CL, it was just returning TRUE if there was at least one
element in the returned enrollment_preparations map. That was wrong,
since GetEnrollmentPreparations returns a map with true/false values
for all known ACA types. So, even when the attestation is not prepared
yet, the map is not empty, just all values are false.

This CL fixes the logic and checks if there is at least one true
value in the map instead.

BUG=b:124085203
TEST=Run "cryptohome --action=tpm_attestation_status" on an unowned
     device with TPM2.0; check that "Attestation Prepared: false".
     Repeat after going through owning and preparing for attestation,
     check that "Attestation Prepared: true" this time.

Change-Id: Iac4c931ba47e4a2754b8f8ad4bf4bebd39f6144a
Reviewed-on: https://chromium-review.googlesource.com/1512750
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Louis Collard <louiscollard@chromium.org>