cryptohome: Add an empty cryptohome-proxy binary to cryptohome
This change adds the cryptohome-proxy binary, which will host the code that
will expose the old DBus interface and forward them to the new DBus interface.
This change setup the environment for the new binary, but does not have any
functionality in it yet.

BUG=b:128271460
TEST=manual && FEATURES=test emerge-eve cryptohome
CQ-DEPEND=CL:1518113

Change-Id: I93dbcd109273937313df22bb9cd72ba1bd4bc36c
Reviewed-on: https://chromium-review.googlesource.com/1517065
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: John  Chen <zuan@chromium.org>
Reviewed-by: John  Chen <zuan@chromium.org>