diagnostics: add status_message field to GetRoutineUpdateResponse.
GetRoutineUpdateResponse has a single field, output, which holds
both the result of a routine, e.g. "Failed to start the urandom
executable" as well as any accumulated logs. Add a new field to
GetRoutineUpdateResponse, status_message, to hold the result of
a routine, and let output be only the accumulated logs. This
allows for a routine like battery_sysfs to separate the log
messages from a detailed status message.

BUG=chromium:919151
TEST=manual and unit tests

Change-Id: I6935d73dcc99ced9fe6ef7a1e6fe72b22b518d94
Signed-off-by: Paul Moy <pmoy@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1516235