cryptohome: Fix D-Bus key challenge delegate object path
Fix the XML config to contain the complete path of the
CryptohomeKeyDelegate D-Bus object - seems that this is required by
chromeos-dbus-bindings.

Without this, the code was crashing when making the D-Bus call - as
the bindings use the value from the XML config to pass it to the D-Bus
library, and the latter asserts the validatity of the path.

BUG=chromium:842791
TEST=manual

Change-Id: Id38640c813a26c33293e805b3b75ceb38a5f3d34
Reviewed-on: https://chromium-review.googlesource.com/1520650
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>