cryptohome: Verify D-Bus key delegate service name
Exit early when the received D-Bus service name (that implements
CryptohomeKeyDelegateInterface) is invalid.

This helps to avoid crashing inside the D-Bus library since it
asserts the validity of the service name. (This problem was already
fixed at the libchrome level in upstream in crbug.com/927196, but the
Chrome OS version of libchrome doesn't have this fix yet.)

BUG=chromium:842791
TEST=manual

Change-Id: I436c584742786c6644d2c5334bdbff8add551d5b
Reviewed-on: https://chromium-review.googlesource.com/1520651
Commit-Ready: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>