camera: intel: ipu3: Use width info not stride info for the BLOB Buffer
The width info from the GBM for the BLOB buffer
is not the same as the stride info.
And the camera framework uses the size
which is the same as the width for the BLOB buffer.
Align the HAL and the camera framework for BLOB buffer size.

BUG=b:127904511
TEST=Build and deploy package to platform and camera works

Change-Id: Iccc644d9c2d2407752dffb3831aa8ab3c31d1c05
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1529727
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wei Meng <wei.meng@intel.com>
Reviewed-by: Hung-yu Wu <hywu@chromium.org>
Reviewed-by: Wei Meng <wei.meng@intel.com>