libbrillo: blkdevutils: avoid passing non-null terminated data as C-string
SecureBlobs created from std::strings do not end in '\0'.
In this case, the dm code calls strlen() on the parameter string
which results in an access beyond the allocated space. This caused
the ASAN builder to fail continuously. Instead, for now, use the
string functions for SecureBlob to pass the data on.

BUG=chromium:941727
TEST=amd-generic vm boots up.

Change-Id: I6c63b37bae1204c9dca6315f2285f16f4405296a
Reviewed-on: https://chromium-review.googlesource.com/1524550
Commit-Ready: Sarthak Kukreti <sarthakkukreti@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>