arc: Run ServerProxy on a dedicated thread.
This is preparation to introduce Fuse daemon.
Fuse will run on the main thread, and instead this CL moves
the thread to run ServerProxy from the main thread to the dedicated
thread.
The detailed implementation is shared with ClientProxy, so the
class name ClientProxyService is renamed to ProxyService.

Also, this adds unittest.

BUG=b:123432308
TEST=Ran cros_run_unit_tests.

Change-Id: I568e610b58832ffba7678a98eaf483c6a35edf4c
Reviewed-on: https://chromium-review.googlesource.com/1526074
Commit-Ready: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>