shill: Change kStatePortal to kStateNoConnectivity
To improve shill portal detection, we plan to add a new service
state indicating that we have received a redirect (HTTP 302) and a
redirect URL from the HTTP probe. This new state is a subset of
the old kStatePortal state, which stood for general lack of
connectivity.

In this change, we transition the kStatePortal state to
kStateNoConnectivity in preparation for the new state.

BUG=chromium:903645
TEST=unittests

Change-Id: I15cf97463124826346541c44ee906c9134f39573
Reviewed-on: https://chromium-review.googlesource.com/1531339
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>