crash: Improved error logging for crash-reporter
This changes some LOG to PLOG where errno should be set. It also
consolidates a RunAndCaptureOutput function so that we can grab stderr
when executing some processes and adds a LogMultilineError so we print
out what happened on stderr. It also consolidates a call to execute
gzip.

BUG=chromium:763516
TEST=Unit tests pass

Change-Id: Idca94c86ff35496b1d580617456c49ecb2365923
Reviewed-on: https://chromium-review.googlesource.com/1534322
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>