cros-disks: Query the storage device path of a device
For each block device, the chromeos disk mount manager keeps track of
the sysfs path of the scsi/mmc device that the block device is a part
of. Currently, it uses scsi/mmc device mount and unmount events to keep
track of all such sysfs paths. However, this does not work for already
mounted devices, so have cros-disks lookup this path instead, and return
it in CrosDisksServer::GetDeviceProperties().

BUG=chromium:942334
TEST=unit tests, platform_CrosDisks.*
TEST=dbus-send --print-reply --type=method_call --system           \
     --dest=org.chromium.CrosDisks /org/chromium/CrosDisks         \
     org.chromium.CrosDisks.GetDeviceProperties string:'/dev/sda1'

Change-Id: I37e5e6b2a9a144aded7a4fef6ea141c0ceb30cbf
Reviewed-on: https://chromium-review.googlesource.com/1540017
Commit-Ready: Austin Tankiang <austinct@chromium.org>
Tested-by: Austin Tankiang <austinct@chromium.org>
Reviewed-by: Anand Mistry <amistry@chromium.org>