runtime_probe: separate the evaluation to different processes
We separate the Eval() implementation into different stages so they can
be done in more than one process, which leads to a better security management
model. Previously, we tied everything in a single process, where the sandbox
is an union of all possible probe functions. It is not obvious on what exactly
the seccomp / capabilities / minijail arguments designated to a probe function
if we scale the number of probe functions.

An illustrated example is from

init
 runtime_probe [1]
    Eval() for probe funtion A [1]
    Eval() for probe funtion B [1]

to

init
 runtime_probe [2]
    Eval() for probe function A [2]
      evaluation under sandbox [2]
      InvokeHelper() [2]     // TODO: fold into debugd in next CL
        runtime_probe_helper (for probe function A) [3]
           EvalInHelper() [3]
      evaluation under sandbox [2]

    Eval() for probe function B [2]
       evaluation under sandbox [2]
       InvokeHelper() [2]     // TODO: fold into debugd in next CL
         runtime_probe_helper (for probe function B) [4]
            EvalInHelper() [4]
       evaluation under sandbox [2]

, where the sandbox surfaces are:
[1]: union of [2], [3], [4]
[2]: communicate to others, for example: mojo, D-Bus
[3]: extended sandbox just for probe function A
[4]: extended sandbox just for probe function B

With this change, we can apply smaller sandboxes instead of a giant sandbox
in the future.

BUG=b:129313193
TEST=USE="cros-debug" emerge-nami runtime_probe
cros_deploy $ACTIVE_DUT runtime_probe
runtime_probe --debug

Cq-Depend: chromium:1545471
Change-Id: Ie25729653012421579eca980fc84d880688a8974
Reviewed-on: https://chromium-review.googlesource.com/1545594
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Ready: Hong-Min Chu <hmchu@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>