vm_tools: concierge: Run crosvm commands synchronously
Currently concierge spawns a process to run a crosvm command but doesn't
wait for it to exit.  This is a problem at suspend time because
concierge will report suspend readiness to powerd and the crosvm command
process will race with the system going to suspend.  If the system
suspends before all vcpus have been stopped the guest kernel will panic
on resume because it thinks there is a soft lockup.

Always wait for the crosvm command process to complete before
continuing.

BUG=chromium:946847
TEST=suspend and resume a device with a VM running.  check the logs to
     see that the "Setting resolv config" message appears in the logs
     after crosvm logs that the run mode has changed to running.

Change-Id: Iba91729e126319c59f030d932003ffff3bd8e549
Reviewed-on: https://chromium-review.googlesource.com/1541848
Commit-Ready: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>