shill: add kStatePortalSuspected state
We plan on adding another state, kStatePortalSuspected, to the state
machine. Shill should transition to this state when it suspects that
it is behind a portal (but hasn't received a redirect URL). More
specifically, it will transition to this state when either 1) the HTTP
probe does not receive a 204 or a well-formatted 302 but the HTTPS
probe succeeds or 2) the HTTP probe does not receive a well-formatted
302 and the HTTPS probe fails.

This change should not affect functionality at all. We add it here
as a child CL to the kStateRedirectFound change for the sake of
consolidating CLs involving new portal states.

BUG=chromium:903645
TEST=unittests

Change-Id: I955ea4fa08d2b3d75bd589cac4f3d9cf7294f0eb
Reviewed-on: https://chromium-review.googlesource.com/1546533
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>