power: Turn off display whenever lid is closed.
Update powerd's InternalBacklightController class to turn
the internal display off while leaving any external displays
on when the lid is closed (in the absense of other
higher-priority events like the system suspending).

Previously, StateController would decide that we were
entering docked mode and notify Daemon, which would notify
the various BacklightController implementations. I think
that this is overly complicated and unnecessary.
StateController still determines that we're in docked mode
(so it can avoid suspending even though the lid is closed),
but BacklightController implementations really only need to
know that the lid is closed in order to do the right thing.

BUG=chromium:945512
TEST=Tested following use cases:
 1. Disable dark resume and test wake on doc works on nocturne
 2. Test device wakes from external hid when in docked.
 3. Close the lid and make sure internal_backlight_controller.cc
    sets the brightness to 0.
 4. Close the lid when external display is connected and make sure
    external display stays on.
 5. Suspend when in docked mode, wake from external hid and make sure
    internal backlight stays off.

Change-Id: I2fab4b11e59850e3fa503036fc21a28b55efbad9
Reviewed-on: https://chromium-review.googlesource.com/1553792
Tested-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Commit-Ready: Ravi Chandra Sadineni <ravisadineni@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ravi Chandra Sadineni <ravisadineni@chromium.org>