power: suspend_stress_test: const factor optimization for memory test
Since the refactor of RTC wakealarm moves the RTC programming to the
setuid helper, the data meaning of the suspend time parameter would be
suspend period rather than a budget time, therefore we are now safe to
remove the workarounds for the demanding memory test.
Such that, it not only simplifies the test script but also gives a
const factor optimization.
i.e.,
A Yorp N4100 8GiB SKU could take 74m41.774s in 100 iterations,
while applying the patch it would reduce the time to 67m41.921s;
Think about the scaling, it would save lots of time on partners'
engineering when it comes to the validation of a new memory part.

Notice the time varies by CPU computation power (memory fill) and the
total size of DRAM (presuming do memory check as much as possible)

BUG=chromium:941846, b:73018133
TEST=do 'time suspend_stress_test -c 100 --memory_check' to benchmark
the consumed time.

Change-Id: Id30f9d43d4ad15b734083428136d2f7e0106ae94
Signed-off-by: Harry Pan <harry.pan@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1547716
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>