u2fd: Store secret and counter as protos
Stores the data inside protos rather than as raw bytes,
and wraps the protos in a container that includes a
hash for integrity checking.

The use of protos allows us to easily extend functionality,
for example if we need to keep track of migrating the
counter from cr50, or storing the key in an alternate
format.

BUG=b:112603199, b:123161715
TEST=deploy to soraka, register, authenticate,
     check files are created, restart u2fd, check files
     are loaded and authentication still works

Change-Id: I3c64a027c0887703af2591368a22ad0df34f2bb9
Reviewed-on: https://chromium-review.googlesource.com/1550190
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Louis Collard <louiscollard@chromium.org>