bluetooth: Set all the properties before exporting a device object
Some clients (e.g. Chrome) relies on complete property values when the
device object is first exported. For example, if a device object is
exported and soon after its Paired property is updated to true, it
thinks that the device has just finished pairing process and shows a
notification about it, while it shouldn't because the device had been
paired previously in the past.

BUG=chromium:840517
TEST=Manual test with Chrome: Pair to a device, restart newblued, and
check that Chrome does not show a notification about a newly paired
device.

Change-Id: I1058c842c60715f29ff4bbfdc54949c0f9897b53
Reviewed-on: https://chromium-review.googlesource.com/1553788
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>