bluetooth: Forward org.bluez.Adapter1.RemoveDevice to BlueZ and NewBlue
This is the other side of CL:1554269. NewBlue should just do no-op and
return true if the device to be removed is already not present.
Additionally configures the dispatcher to forward RemoveDevice to both
BlueZ and NewBlue.

BUG=chromium:840517
TEST=Use bluetoothctl to connect to btdispatch, do a scan and try to
remove a Classic device and an LE device. The operations should succeed.

Change-Id: I2f4b5ebef63d909075a1d1e21f58bbac84dc4c8b
Reviewed-on: https://chromium-review.googlesource.com/1554223
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>