cryptohome: Remove precomputed salt challenge-response signature
Remove the codepath that allowed to pass precomputed salt signature
to the challenge-response vault keyset decryption operation.

For the background, this codepath was originally added to allow for
an optimization of the vault keyset decryption operation immediately
after its creation - so that we're doing only one challenge-response
for decrypting the TPM-protected secret, and reuse the salt's
signature obtained during the creation step.

This codepath won't be useful anymore, since the recent changes in
the encryption algorithms (e.g., crrev.com/c/1524781) allow not
doing a decryption operation when the vault keyset is created.

BUG=chromium:806788
TEST=existing unit tests ChallengeCredentialsDecryptOperation*Test

Change-Id: I20ade9d535f55ca33399a3da36a4da194250e6fd
Reviewed-on: https://chromium-review.googlesource.com/1557814
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>